requirements.txt
streamlit==1.34.0
langchain==0.2.1
langchain-community==0.2.1
chromadb==0.4.24
pypdf==5.3.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.32.3